refactor: rename airbyte/hooks
